// src/shared/types.ts
export interface WindowState {
  width: number
  height: number
  x?: number
  y?: number
  isMaximized: boolean
}